VPA: Stuck in-place resizes still require eviction
So because of InPlacePodVerticalScaling, we can have a pod object whose
resource spec is correct, but whose status is not, because that pod may
have been updated in-place after the original admission.

This would have been ignored until now because "the spec looks correct",
but we need to take the status into account as well if a resize is in
progress.

This commit:
- takes status resources into account for pods/containers that are being
  in-place resized
- makes sure that any pods that are "stuck" in-place updating (i.e. the
node doesn't have enough resources either temporarily or permanently)
will still show up in the list as having "wrong" resources so they can
still get queued for eviction and be re-assigned to nodes
that do have enough resources